+++ lib/Sema/CSSolver.cpp
@@ -1985,6 +1985,13 @@ ConstraintSystem::solve(Expr *&expr,
 
   assert(!solverState && "use solveRec for recursive calls");
 
+  // If the frontend options for dumping expression times or warning
+  // about long expression type checking are enabled, set up the
+  // timer.
+  if (TC.getDebugTimeExpressions() || TC.getWarnLongExpressionTypeChecking())
+    Timer.emplace(expr, TC.getDebugTimeExpressions(),
+                  TC.getWarnLongExpressionTypeChecking(), TC.Context);
+
   // Try to shrink the system by reducing disjunction domains. This
   // goes through every sub-expression and generate its own sub-system, to
   // try to reduce the domains of those subexpressions.
